refactor: move file validation
